Safari-609.1.11.1/LayoutTests/http/tests/security/xss-DENIED-assign-location-hash-expected.txt

CONSOLE MESSAGE: line 4: SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.

PASS: cross-site assignment of location.hash not allowed
